% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{orbital_mutate}
\alias{orbital_mutate}
\title{Use orbital in a mutate way}
\usage{
orbital_mutate(.data, x, only_pred = FALSE)
}
\arguments{
\item{.data}{A data frame that can be used with mutate.}

\item{x}{A orbital object.}

\item{only_pred}{A logical value to determine if whole data set should be
returned or just predictions.}
}
\value{
A modified data frame.
}
\description{
Use orbital in a mutate way
}
\examples{
library(workflows)
library(recipes)
library(parsnip)

rec_spec <- recipe(mpg ~ ., data = mtcars) \%>\%
  step_normalize(all_numeric_predictors())

lm_spec <- linear_reg()

wf_spec <- workflow(rec_spec, lm_spec)

wf_fit <- fit(wf_spec, mtcars)

orbital_obj <- orbital(wf_fit)

mtcars \%>\%
  orbital_mutate(orbital_obj)

mtcars \%>\%
  orbital_mutate(orbital_obj, only_pred = TRUE)
}
