% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R, R/predict.R
\name{orbital_predict}
\alias{orbital_predict}
\title{Use orbital in a mutate way}
\usage{
orbital_predict(.data, x)

orbital_predict(.data, x)
}
\arguments{
\item{.data}{A data frame that can be used with mutate.}

\item{x}{A orbital object.}
}
\value{
A modified data frame.

A modified data frame.
}
\description{
Use orbital in a mutate way

Use orbital in a mutate way
}
\examples{
library(workflows)
library(recipes)
library(parsnip)

rec_spec <- recipe(mpg ~ ., data = mtcars) \%>\%
  step_normalize(all_numeric_predictors())

lm_spec <- linear_reg()

wf_spec <- workflow(rec_spec, lm_spec)

wf_fit <- fit(wf_spec, mtcars)

orbital_obj <- orbital(wf_fit)

mtcars \%>\%
  orbital_predict(orbital_obj)
library(workflows)
library(recipes)
library(parsnip)

rec_spec <- recipe(mpg ~ ., data = mtcars) \%>\%
  step_normalize(all_numeric_predictors())

lm_spec <- linear_reg()

wf_spec <- workflow(rec_spec, lm_spec)

wf_fit <- fit(wf_spec, mtcars)

orbital_obj <- orbital(wf_fit)

mtcars \%>\%
  orbital_predict(orbital_obj)
}
