% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.R
\name{to_r_fun}
\alias{to_r_fun}
\title{Turn weasel object into a R function}
\usage{
to_r_fun(x, name = "weasel_predict", file)
}
\arguments{
\item{x}{A weasel object.}

\item{name}{Name of created function. Defaults to `"weasel_predict"``.}

\item{file}{A file name.}
}
\value{
A weasel object.
}
\description{
Turn weasel object into a R function
}
\examples{
library(workflows)
library(recipes)
library(parsnip)

rec_spec <- recipe(mpg ~ ., data = mtcars) \%>\%
  step_normalize(all_numeric_predictors())

lm_spec <- linear_reg()

wf_spec <- workflow(rec_spec, lm_spec)

wf_fit <- fit(wf_spec, mtcars)

weasel_obj <- weasel(wf_fit)

file_name <- tempfile()

to_r_fun(weasel_obj, file = file_name)

readLines(file_name)
}
