% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weasel.R
\name{weasel}
\alias{weasel}
\title{Turn tidymodels workflows into equations}
\usage{
weasel(x, ...)
}
\arguments{
\item{x}{A workflow object.}

\item{...}{Not currently used.}
}
\value{
A weasel object.
}
\description{
Turn tidymodels workflows into equations
}
\examples{
library(workflows)
library(recipes)
library(parsnip)

rec_spec <- recipe(mpg ~ ., data = mtcars) \%>\%
  step_normalize(all_numeric_predictors())

lm_spec <- linear_reg()

wf_spec <- workflow(rec_spec, lm_spec)

wf_fit <- fit(wf_spec, mtcars)

weasel(wf_fit)

}
