% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{weasel_dt}
\alias{weasel_dt}
\title{Convert to data.table code}
\usage{
weasel_dt(x)
}
\arguments{
\item{x}{A weasel object.}
}
\value{
data.table code.
}
\description{
Convert to data.table code
}
\examples{
library(workflows)
library(recipes)
library(parsnip)

rec_spec <- recipe(mpg ~ ., data = mtcars) \%>\%
  step_normalize(all_numeric_predictors())

lm_spec <- linear_reg()

wf_spec <- workflow(rec_spec, lm_spec)

wf_fit <- fit(wf_spec, mtcars)

weasel_obj <- weasel(wf_fit)

weasel_dt(weasel_obj)
}
