% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{weasel_inline}
\alias{weasel_inline}
\title{Use weasel object splicing function}
\usage{
weasel_inline(x)
}
\arguments{
\item{x}{A weasel object.}
}
\value{
a list of quosures.
}
\description{
Use weasel object splicing function
}
\details{
This function is mostly going to be used for
\href{https://rlang.r-lib.org/reference/topic-inject.html#dots-injection}{Dots Injection}.
See examples for use cases.
}
\examples{
library(workflows)
library(recipes)
library(parsnip)

rec_spec <- recipe(mpg ~ ., data = mtcars) \%>\%
  step_normalize(all_numeric_predictors())

lm_spec <- linear_reg()

wf_spec <- workflow(rec_spec, lm_spec)

wf_fit <- fit(wf_spec, mtcars)

weasel_obj <- weasel(wf_fit)

library(dplyr)

mtcars \%>\%
  mutate(!!!weasel_inline(weasel_obj))
}
