% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{weasel_json_read}
\alias{weasel_json_read}
\title{Read weasel json file}
\usage{
weasel_json_read(path)
}
\arguments{
\item{path}{file on disk.}
}
\value{
A weasel object
}
\description{
Read weasel json file
}
\examples{
library(workflows)
library(recipes)
library(parsnip)

rec_spec <- recipe(mpg ~ ., data = mtcars) \%>\%
  step_normalize(all_numeric_predictors())

lm_spec <- linear_reg()

wf_spec <- workflow(rec_spec, lm_spec)

wf_fit <- fit(wf_spec, mtcars)

weasel_obj <- weasel(wf_fit)

tmp_file <- tempfile()

weasel_json_write(weasel_obj, tmp_file)

weasel_json_read(tmp_file)
}
\seealso{
\code{\link[=weasel_json_write]{weasel_json_write()}}
}
