% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{weasel_sql}
\alias{weasel_sql}
\title{Convert to SQL code}
\usage{
weasel_sql(x, con)
}
\arguments{
\item{x}{A weasel object.}

\item{con}{A connection object.}
}
\value{
SQL code.
}
\description{
Convert to SQL code
}
\examples{
library(workflows)
library(recipes)
library(parsnip)

rec_spec <- recipe(mpg ~ ., data = mtcars) \%>\%
  step_normalize(all_numeric_predictors())

lm_spec <- linear_reg()

wf_spec <- workflow(rec_spec, lm_spec)

wf_fit <- fit(wf_spec, mtcars)

weasel_obj <- weasel(wf_fit)

library(dbplyr)
con <- simulate_dbi()

weasel_sql(weasel_obj, con)
}
