% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{orbital_inline}
\alias{orbital_inline}
\title{Use orbital object splicing function}
\usage{
orbital_inline(x)
}
\arguments{
\item{x}{A orbital object.}
}
\value{
a list of quosures.
}
\description{
Use orbital object splicing function
}
\details{
This function is mostly going to be used for
\href{https://rlang.r-lib.org/reference/topic-inject.html#dots-injection}{Dots Injection}.
See examples for use cases.
}
\examples{
\dontshow{if (rlang::is_installed("recipes")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(workflows)
library(recipes)
library(parsnip)

rec_spec <- recipe(mpg ~ ., data = mtcars) \%>\%
  step_normalize(all_numeric_predictors())

lm_spec <- linear_reg()

wf_spec <- workflow(rec_spec, lm_spec)

wf_fit <- fit(wf_spec, mtcars)

orbital_obj <- orbital(wf_fit)

library(dplyr)

mtcars \%>\%
  mutate(!!!orbital_inline(orbital_obj))
\dontshow{\}) # examplesIf}
}
